+++ core/src/com/astetyne/expirium/server/net/ServerPacketManager.java
@@ -185,7 +185,7 @@ public class ServerPacketManager {
 
     public void putEnviroPacket() {
         out.startPacket(28);
-        out.putInt(GameServer.get().getWorld().getWorldTime());
+        out.putFloat(GameServer.get().getWorld().getWorldTime());
         out.putInt(GameServer.get().getWorld().getWeather().getID());
     }
